infrastructure/requirements.txt
aws-cdk.aws-lambda
aws-cdk.aws-apigateway
aws-cdk.core

